Improve task abstractness

+++ Node/Tasks/Exec/Vectorizer.cs
@@ -11,30 +11,29 @@ public static class VectorizerTasks
     public static IEnumerable<IPluginAction> CreateTasks() => new IPluginAction[] { new VeeeVectorize() };
 
 
-    class VeeeVectorize : PluginAction<VeeeVectorizeInfo>
+    class VeeeVectorize : InputOutputPluginAction<VeeeVectorizeInfo>
     {
         public override string Name => "VeeeVectorize";
         public override PluginType Type => PluginType.VeeeVectorizer;
         public override FileFormat FileFormat => FileFormat.Jpeg;
 
-        protected override async Task<string> Execute(ReceivedTask task, VeeeVectorizeInfo data)
+        protected override async Task Execute(ReceivedTask task, VeeeVectorizeInfo data, ITaskInput input, ITaskOutput output)
         {
-            task.InputFile.ThrowIfNull();
-            var output = task.FSOutputFile();
+            var outputfile = task.FSOutputFile();
 
             var exepath = task.GetPlugin().GetInstance().Path;
             var args = task.InputFile;
 
             var plugindir = Path.GetDirectoryName(exepath)!;
             var outdir = Path.Combine(plugindir, "out");
-            if (Directory.Exists(outdir))
-                Directory.Delete(outdir, true);
-            Directory.CreateDirectory(outdir);
 
+            // assuming we aren't executing several of veee.exe simultaneously
+            if (Directory.Exists(outdir)) Directory.Delete(outdir, true);
+            Directory.CreateDirectory(outdir);
             File.WriteAllText(Path.Combine(plugindir, "config.xml"), GetConfig(data.Lods));
 
             await ExecuteProcess(exepath, args, false, delegate { }, task);
-            return output;
+            await UploadResult(task, output, outputfile);
         }
